refactored for better testing coverage
